check cvssrating and move deprecate to json api
