+++ src/NuGetGallery/App_Start/Routes.cs
@@ -241,11 +241,6 @@ namespace NuGetGallery
                 "packages/manage/delete",
                 new { controller = "Packages", action = "Delete" });
 
-            var packageDeprecateRoute = routes.MapRoute(
-                RouteName.PackageDeprecateAction,
-                "packages/manage/deprecate",
-                new { controller = "Packages", action = nameof(PackagesController.Deprecate) });
-
             var confirmationRequiredRoute = routes.MapRoute(
                 "ConfirmationRequired",
                 "account/ConfirmationRequired",
